edge-middleware/requirements/requirements.txt
sqlalchemy
marshmallow_sqlalchemy
paho-mqtt
